update deps & make use of node 20 instead of node 16
