+++ action.yml
@@ -18,5 +18,5 @@ inputs:
     required: false
     default: "UTC"
 runs:
-  using: "node16"
-  main: "dist/index.js"
+  using: "node20"
+  main: "dist/index.mjs"
