+++ package.json
@@ -1,11 +1,11 @@
 {
   "name": "set-timezone",
-  "version": "1.2",
+  "version": "2.0",
   "description": "Set timezone",
   "main": "index.js",
   "scripts": {
     "run": "node index.js",
-    "build": "ncc build index.js",
+    "build": "ncc build index.mjs",
     "test": "jest"
   },
   "repository": {
@@ -19,11 +19,14 @@
   },
   "homepage": "https://github.com/szenius/set-timezone#readme",
   "dependencies": {
-    "@actions/core": "^1.2.5",
-    "execa": "^4.0.3"
+    "@actions/core": "^1.10.1",
+    "execa": "^8.0.1"
   },
   "devDependencies": {
-    "@zeit/ncc": "^0.22.3",
-    "jest": "^27.0.6"
+    "@vercel/ncc": "^0.38.1",
+    "jest": "^29.7.0"
+  },
+  "engines": {
+    "node": "20"
   }
 }
